Modified Bicycle class: Introduced inheritance by creating MountainBike as subclass of Bicycle
The style variable effectively divides instances of Bicycle into two different kinds of things. These two things share a great deal of behavior but differ along the style dimension. This single class contains several different, but related, types. This is the exact problem that inheritance solves. Another class MountainBike is created which acts as subclass of Bicycle and has its own constructor and overrides getSpares. In MountainBike constructor the super class constructore is called first.

Bicycle class:
	Member variables:
		Removed: style, rear_shock, front_shock
	Methods:
		Modified: Bicycle - initializes only size, tape_color and spares
			  getSpares - removed style bases if else condition

MountainBike class : extends Bicycle class
	Member variables:
		Added: rear_shock, front_shock
	Methods:
		Added: MountainBike - calls super class constructor and initialize rear_shocka nd front_shock
		       getSpares - adds rear_shock spare in spares of super class

Issues: The mountainbike still has member variables such as tire_size which are inherited from Bicycle class but not needed in MountainBike class. The Bicycle class contains behavior that is appropriate for both a peer and a parent of MountainBike. Some of the behavior in Bicycle is correct for MountainBike, some is wrong, and some doesn’t even apply. As written, Bicycle should not act as the superclass of MountainBike.