Added spares field in Bicycle class: Implemented spares in sub-classes
Added spares field in Bicycle class in which getSpares function is defined which add the tire_size and chain fields which are common for all subclassses. In the sublass definition of getSpares, at first getSpares of super is called and the additional fields such as tape_color in case of RoadBike and rear_shock in case of MaountainBike are added to spares respectively. A pattern in the design can be observed - Every template method sent by Bicycle is implemented in Bicycle itself, and MountainBike and RoadBike both send super in their constructor and spares methods.

Apart from this, methods are implemented to get default value of tire_size and chain depending upon the class of the object. These methods are overriden in sub-classes as per the requirement.

Bicycle:
	Member variables:
		Added: spares
	Methods:
		Modified: Bicycle - calls functions to get respective default values for chain and tire_size, initialize spares using getSpares
		Added: getSpares - assign tire_size and chain fields
		       getDefaultChain - return default value of chain
		       getDefaultTire - return default value of tire_size

RoadBike:
	Methods:
		Added: getSpares - calls super class getSpares and adds tape_color
		       getDefaultTire

MountainBike:
	Methods:
		Added: getSpares - calls super class getSpares and adds rear_shock
		       getDefaultTire

Issues/ToDo - the MountainBike and RoadBike subclasses follow a similar pattern. They each know things about themselves and things about their superclas. Knowing things about other classes creates dependencies and dependencies couple objects together. When RecumbentBike fails to send super during initialize it misses out on the common initialization provided by Bicycle and does not get a valid size, chain, or tire size. This error can manifest at a time and place far distant from its cause, making it very hard to debug.