Final implementation: Pending issues due to inheritance
As mentioned in previous commit, all the sub-classes are by-default types of parent class. So there is an inherent dependency between the parent and child class. Everytime an instance of subclass is created, the constructor of super class is always called. This leads to design issues as discussed previously and does not follow SOLID principles.

So a better practice in this scenario would be use composition instead of inheritance.