Implemented Composition on Bicycle:
Removed inheritance from the implementation. Now the class Bicycle contains two mamber variables, size and parts where parts is an instance of class Parts. Parts class contains a set of Part type objects which represent each part of bicycle. Each part object contains three values, name of part, value of part and a boolean value indicating whether the part is a spare. The static method build of class PartsFactory creates an object of Parts type using the config of the bicycle. In this way, we can create any type of bicycle using the bicycle_config and using it to create Parts object which initializes the instance of a Bicycle. The getSpares is implemented by returning all the parts which have the vale of needSpare set as true.